Removed unit tests for the HelpDisplay component, instead opting to only use integration tests as per the react testing library philosophy. Completed tests for the HelpDisplay component
